Package Updates (#107)
* Migrates to the new ESLint configuration format

Replaces the deprecated `.eslintrc.js` and `.eslintignore` files with the new `eslint.config.js` format.

Updates dependencies to align with the new ESLint flat config.
This ensures compatibility with the latest ESLint version and enables modern linting features.

* Updated jest and types/jest

* Updates @types/node dependency

Updates the @types/node dependency to the latest version.

This ensures that the project benefits from the latest type definitions
and improvements for Node.js development.

* Updates Node.js and npm versions

Updates the Node.js version in the Dockerfile and package.json to a newer LTS release.

Also updates npm to the latest available version, ensuring compatibility and access to the latest features and security patches.

This upgrade ensures the project benefits from the performance improvements, bug fixes, and security updates included in the newer runtime and package manager versions.