Update packages (#115)
* Updates dependencies in package-lock.json

Updates various dependencies in package-lock.json and package.json to address compatibility issues and incorporate new features. This includes updates to eslint-plugin-prettier, supertest, and other packages to align with the latest project requirements and to improve overall stability.

* Updates redis to version 5.8.2

Updates the redis dependency to the latest version.

This upgrade includes bug fixes, performance improvements, and new features.

* Update travisCI node version

* Update typescript